+++ api/app.js
@@ -2,6 +2,7 @@ import express from 'express'
 import path from 'path'
 import userRoutes from './routes/userRoutes.js'
 import cookieParser from "cookie-parser";
+import tokenRoutes from './routes/tokenRoutes.js';
 
 const app = express()
 
@@ -11,6 +12,7 @@ app.use(express.static(path.resolve('public')))
 
 // rotas da API
 app.use('/users', userRoutes)
+app.use('/tokens', tokenRoutes)
 
 app.get('/', (req, res) => {
   res.sendFile(path.resolve('public', 'index.html'))