+++ api/routes/userRoutes.js
@@ -1,15 +1,15 @@
 import { Router } from 'express'
 import { UserController } from '../controllers/userController.js'
-import { cookieAuth } from '../middleware/cookieAuth.js'
+import { authorizeToken } from '../middleware/authToken.js'
 
 const userRoutes = Router()
 
 userRoutes.post('/', UserController.create)
-userRoutes.post('/login', cookieAuth, UserController.login)
-userRoutes.get('/', cookieAuth, UserController.list)
-userRoutes.get('/:id', cookieAuth, UserController.getOne)
-userRoutes.post('/atualizar', cookieAuth, UserController.updateMultiple)
-userRoutes.post('/deletar', cookieAuth, UserController.deleteMultiple)
+userRoutes.post('/login', UserController.login)
+userRoutes.get('/', authorizeToken, UserController.list)
+userRoutes.get('/:id', authorizeToken, UserController.getOne)
+userRoutes.post('/atualizar', authorizeToken, UserController.updateMultiple)
+userRoutes.post('/deletar', authorizeToken, UserController.deleteMultiple)
 // userRoutes.delete('/:id', cookieAuth, UserController.delete)
 
 export default userRoutes